Much better fix for the infinite scroll bug
This fix is much better than the previous one which was pretty badly coded and was very unreliable. With this fix, you don't even need to scroll down to trigger the infinite scroll, the script already does that automatically and stops auto scrolling when not needed anymore. It just checks if the scrollbar is visible, if it isn't, then it creates a tall element relative to the window size, auto scroll to the bottom, then removes it. This is repeated until the list has enough items to make the scroll bar to appear.