New export/import feat.; Custom events (code); Bug fixes
Settings page:
-- Implemented the ability to export/import a config's settings. Two buttons were added in the Settings page, one to export and another to import.   
All settings and entries from the Security, Privacy, Parental Control, Denylist, Allowlist and Setttings pages are exported in a .json file named after the exported config id. You can freely rename the file. The file will appear in your browser like a downloaded file, so what exactly happens will depend on your browser's settings for downloads.    
When you import a config, the settings from the file are *merged* with the current config settings, it doesn't *replace* them. If you want to replace the config, you should remove the old settings, or create a new config, and then import from the file.   
Any items that you normally add by clicking an "Add" button won't be imported in the exact same order they were exported, because NextDNS doesn't accept a group of items to be added in one go, so every item need to be added one by one, this means that for every item, a new connection need to be made, and each connection finishes in it's own time, which makes the first connection not necessarily be the first to finish. I plan to add an option to preserve the same order, but keep in mind that enabling such option would make the importing take much longer, specially for configs with lots of entries.    
It can happen that an item or two aren't imported because of a network or timing issue. In that case, you can manually add the missing items or import again. You can reimport any number of times, no items are duplicated by doing this.
Currently you can only export and import settings from all pages. I plan adding some customization options. In the meantime, you can comment out the code that imports/exports the pages you don't want.  

Security page:
-- Improved the "Add all TLDs" button behavior. Now it alerts the user if there's no more TLD left to be added; And fixed bug which made the page not be reloaded when finished adding the TLDs;

Allowlist/Denylist pages:
-- Fixed bug that caused the allow/denylist domains to get "squeezed" when enabling the "bold root domain" option;

Code:
-- Changed the code to use custom events for requests, instead of strings and eval;
-- Simplified the code of the fix for the infinite scroll bug. Now it just changes the body's margin-bottom instead of creating an element;
-- Done some optimizations and cleanings in the code.