Reverted back to requiring permission for api.nextdns.io by default
I changed the permission for api.nextdns.io to be optional, as only a few browsers actually require it, but by doing so, the browser never asks for such permission, so not including in the manifest and including as optional has basically the same effect. To actually have the permission granted, it needs to be asked manually by the extension itself. So to avoid unnecessary complexities, api.nextdns.io is now required by default again (just like 4.5.1).