Removed the uninstallation feedback
Removed the uninstallation feedback. Only a small minority of users uninstalling actually answered it. Chrome Web Store says that there was 165 uninstallations since it was published (roughly, and without considering other browsers). The total of responses is 35. Out of these 35 responses, most of them:

1. complain about something that you can't know what the heck they are talking about because of the vagueness in their "explanation"; 
2. complain about something that doesn't even make sense; 
3. vaguely complain about something without mentioning the browser;
4. mention the browser but doesn't mention what is not working;
5. say that something is not working but leave the explanation box empty;
8. invalid responses (i.e. responses that are well-intentioned but serve no purpose);
6. fill in with random characters;
7. troll(s).

Excluding all these responses, it remains only three, which were actually useful. But it makes no sense to keep this form to only less than 10% of responses be useful and actually intended to help.